requirements.txt
cython
tqdm
msgpack
pillow
numpy
scipy
scikit-image
gdown
torchmetrics
